Fix error with first-time sync and speed up sync command
Previously the sync failed when the device-specific manifest wasn't downloaded yet, because some local manifests attempted to override projects that didn't exist yet. An added check for some of those device-specific files fixes this problem. I've also removed the -d option on the sync command because it isn't necessary and causes slowdown.